Scaffold example fixes and removed loadtest subcommand.
